requirements.txt
aiofiles==24.1.0
amqp==5.2.0
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
billiard==4.2.0
bleach==6.1.0
blis==0.7.11
build==1.0.3
bx_py_utils==93
CacheControl==0.14.0
catalogue==2.0.10
celery==5.4.0
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
ci-info==0.3.0
cleo==2.1.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cloudpathlib==0.18.1
confection==0.1.5
configobj==5.0.8
configparser==7.0.0
crashtest==0.4.1
cron-descriptor==1.4.3
cryptography==41.0.4
cymem==2.0.8
distlib==0.3.7
distro==1.9.0
Django==5.0.7
django-allauth==0.63.6
-e git+https://github.com/celery/django-celery-beat.git#egg=django-celery-beat
#torch
#torchvision
#torchaudio
django-celery-results==2.5.1
django-environ==0.11.2
django-jazzmin==3.0.0
django-redis==5.4.0
django-rest-framework==0.1.0
django-timezone-field==7.0
djangorestframework==3.15.2
dulwich==0.21.6
etelemetry==0.3.1
fastjsonschema==2.18.1
filelock==3.12.2
frontend==0.0.3
fsspec==2024.2.0
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
huggingface-hub==0.24.2
icdiff==2.0.7
idna==3.4
importlib-metadata==6.8.0
installer==0.7.0
isodate==0.6.1
itsdangerous==2.2.0
jaraco.classes==3.3.0
jeepney==0.8.0
Jinja2==3.1.3
joblib==1.4.2
keyring==24.2.0
kombu==5.3.7
langcodes==3.4.0
language_data==1.2.0
looseversion==1.3.0
lxml==5.2.2
marisa-trie==1.2.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
more-itertools==10.1.0
mpmath==1.3.0
msgpack==1.0.7
murmurhash==1.0.10
networkx==3.2.1
nibabel==5.2.1
nipype==1.8.6
nltk==3.8.1
numpy==1.26.3
openai==1.37.1
packaging==23.2
pandas==2.2.2
pathlib==1.0.1
pexpect==4.9.0
pgvector==0.3.2
pillow==10.2.0
pipenv==2024.0.1
pkginfo==1.10.0
pl-core-news-sm @ https://github.com/explosion/spacy-models/releases/download/pl_core_news_sm-3.7.0/pl_core_news_sm-3.7.0-py3-none-any.whl#sha256=94fd294e49f87fccf863f2792a1d89be1a0fd0a3c549790b2c8338ac304bee36
platformdirs==3.11.0
poetry==1.8.3
poetry-core==1.9.0
poetry-plugin-export==1.8.0
pprintpp==0.4.0
preshed==3.0.9
prompt_toolkit==3.0.47
prov==2.0.1
psycopg==3.2.1
psycopg-binary==3.2.1
ptyprocess==0.7.0
pycparser==2.21
pydantic==2.8.2
pydantic_core==2.20.1
pydot==3.0.1
Pygments==2.18.0
PyJWT==2.8.0
PyMuPDF==1.24.9
PyMuPDFb==1.24.9
pypandoc==1.13
pyparsing==3.1.2
pypng==0.20220715.0
pyproject_hooks==1.0.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-docx==1.1.2
pytils==0.4.1
pytz==2024.1
pyxnat==1.6.2
PyYAML==6.0.1
qrcode==7.4.2
rapidfuzz==3.5.2
rdflib==6.3.2
redis==5.0.7
regex==2024.7.24
requests==2.31.0
requests-toolbelt==1.0.0
rich==13.7.1
safetensors==0.4.3
scikit-learn==1.5.1
scipy==1.14.0
SecretStorage==3.3.3
sentence-transformers==3.0.1
shellingham==1.5.3
simplejson==3.19.2
six==1.16.0
smart-open==7.0.4
sniffio==1.3.1
spacy==3.7.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqlparse==0.5.1
srsly==2.4.8
starlette==0.38.1
sympy==1.12
thinc==8.2.5
threadpoolctl==3.5.0
tokenizers==0.19.1
tomlkit==0.12.1
tools==0.1.9
torch==2.4.0
torchaudio==2.4.0
torchvision==0.19.0
tqdm==4.66.4
traits==6.3.2
transformers==4.43.2
trove-classifiers==2023.9.19
typer==0.12.3
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.0.6
uvicorn==0.30.3
vine==5.1.0
virtualenv==20.26.3
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.1
webencodings==0.5.1
wrapt==1.16.0
zipp==3.17.0
